consolidated doesn't really make sense on its own as a module
