+++ ceilometer_publisher_vaultaire/process.py
@@ -5,7 +5,6 @@ from dateutil.tz import tzutc
 from marquise import Marquise
 
 import ceilometer_publisher_vaultaire.siphash
-import ceilometer_publisher_vaultaire.consolidated as consolidated
 
 KEYS_TO_DELETE = [
     "timestamp",
@@ -16,8 +15,7 @@ KEYS_TO_DELETE = [
     "size",
 ]
 
-# FIXME: refactor the functions in this module in such a way that they
-# can log.
+RAW_PAYLOAD_IP_ALLOC = 1
 
 def process_sample(sample):
     processed = []
@@ -146,9 +144,9 @@ def process_consolidated_pollster(sample):
     if name.startswith("instance"):
         payload = hash_flavor_id(metadata["instance_type"])
     elif name.startswith("volume.size"):
-        payload = consolidated.volumeToRawPayload(sanitize(sample["volume"]))
+        payload = volume_to_raw_payload(sanitize(sample["volume"]))
     elif name.startswith("ip.floating"):
-        payload = consolidated.RAW_PAYLOAD_IP_ALLOC
+        payload = RAW_PAYLOAD_IP_ALLOC
     else:
         # We don't need to write out a consolidated version of anything
         # else.
@@ -194,11 +192,11 @@ def process_consolidated_event(sample):
 
     # We use the instance_type_id for instance events
     if name.startswith("instance"):
-        payload = consolidated.construct_payload(metadata["event_type"], metadata.get("message",""), metadata["instance_type_id"])
+        payload = construct_payload(metadata["event_type"], metadata.get("message",""), metadata["instance_type_id"])
     elif name.startswith("volume.size"):
-        payload = consolidated.construct_payload(metadata["event_type"], metadata.get("status",""), consolidated.volumeToRawPayload(sample["volume"]))
+        payload = construct_payload(metadata["event_type"], metadata.get("status",""), volume_to_raw_payload(sample["volume"]))
     elif name.startswith("ip.floating"):
-        payload = consolidated.construct_payload(metadata["event_type"], "", 1)
+        payload = construct_payload(metadata["event_type"], "", 1)
     else:
         payload = sanitize(sample["volume"])
 
@@ -323,3 +321,53 @@ def get_flavor_type(sample):
     elif "instance_type" in sample:
         flavor_type = sample["instance_type"]
     return flavor_type
+
+
+def construct_payload(event_type, message, rawPayload):
+    """
+    eventResolution is passed in message, if none is given assumed to be Success
+    eventResolution takes up the LSByte
+    eventVerb is the action of the event, e.g. create, delete, shutdown, etc.
+    eventVerb takes up the 2nd LSByte
+    eventEndpoint defines whether the data represents an event start, end or an
+    instance event.
+    eventVerb takes up the 3rd LSByte
+    rawPayload takes up the 4 MSBytes (32 bits)
+    rawPayload is specific on counter
+    Start events are even, end are odd (LSB)
+    Other 7 bits for create, update, end, and any future additions
+    """
+
+    # If no event endpoint is included, assume instantaneous event (e.g. image deletion)
+    eventEndpoint = 0
+
+    if message == "Success":
+        eventResolution = 0
+    # Empty message implies success
+    elif message == "":
+        eventResolution = 0
+    elif message == "Failure":
+        eventResolution = 1
+    elif message == "error":
+        eventResolution = 2
+
+    rest,maybeVerb = event_type.rsplit('.', 1)
+    if maybeVerb in ("start", "end"):
+        eventEndpoint = {"start":1, "end":2}.get(maybeVerb)
+        _,verb = rest.rsplit('.', 1)
+    else:
+        verb = maybeVerb
+
+    eventVerb = {"create":1, "update":2, "delete":3, "shutdown":4, "exists":5}.get(verb)
+
+    if eventResolution is None:
+        raise Exception("Unsupported message given to eventToByte")
+    if eventVerb is None:
+        raise Exception("Unsupported event class given to eventToByte")
+    if eventEndpoint is None:
+        raise Exception("Unsupported event endpoint given to eventToByte")
+
+    return eventResolution + (eventVerb << 8) + (eventEndpoint << 16) + (rawPayload << 32)
+
+def volume_to_raw_payload(volume):
+    return volume
